EE-7402 Updated BDD scenarios for name-matching to reflect that the service will try any surname that contains multiple words as the entire name. Updating the tests showed that there were duplicate permutations of the names being used so this was fixed. Also the test class for NameMatchingGenerator was moved to live in the namematching package which is where NameMatchingGenerator lives
